chore(actions): move saveAsActionLogic out of products/actions
products/actions/ isn't a bootstrapped product, so the logic and its
tests don't belong under that tree. Move them to frontend/src/scenes/
actions/ (alongside the existing actionActivityDescriber) and revert
the package.json + lockfile entries this previously needed.

Kea path updates from
['products', 'actions', 'frontend', 'logics', 'saveAsActionLogic']
to
['scenes', 'actions', 'saveAsActionLogic'].

Generated-By: PostHog Code
Task-Id: 5cc6267b-3241-42ec-b029-997cd288ba9e
